requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Jinja2==2.11.2
docx-mailmerge
xlrd==1.1.0
pandas
Flask-Cors
money